Use a pre-pass to emit variables used in the scope
Currently we are emittng alloca's for variables in their own basic block.
This is OK but forces us to jump back and forth the current basic block
and the allocating basic block, which complicates unnecessarily some cases.

It is easier to do a first traversal to gather variables that will have to be
emitted. This avoid having to keep a basic block around only for allocations.

The implemented approach is not as efficient as it could be, for it will
traverse many subtrees many times if there are many lexical scopes. However
this is rare in Fortran. A better approach can do a single prepass per function
and map each symbol to be emitted with its scope. Once the scope is emitted
also emit its variables (if this seems a bit silly it is because all symbols in
a scope may not be used at all, so we do not want to emit every symbol inside a
scope).

Kudos to Sergi for this suggestion.