Fix use after-free
In tl-analysis-check-phase.cpp we see (and similarly for get_ub)

  const Nodecl::NodeclBase& assert_lb = *(c_it->get_lb().begin());

but unfortunately this fails to work as get_lb and get_ub are defined like
this

  NodeclSet InductionVar::get_ub() const

Here the set returned is destroyed at the end of the initialization so its
internal element (accessible by begin()) is destroyed as well.

As copying the set may not make sense, just return a const reference to
it, so its internal object will survive as no copy happens.