Create a single libmercurium library
We're currently creating a big number of shared libraries that are not
phases per se. As they do not have to be dynamically loaded we do not
need them in a shared library, so we could just link everything
statically into the driver.

Doing this, though, would introduce a few difficulties: phases should in
practice be linked against the driver which is conceptually arguable. It
makes more sense to link them again the code of Mercurium itself.

So instead of linking everything into the driver, a sensible trade-off
is creating a libmercurium library which is used by the drivers and the
plugins.

The ultimate goal is to have a clear structure where we have library
code of the compiler isolated from the driver itself (and if needed
provide a library into Mercurium that provides driver-like support).
Currently we're still using some functions in the driver (e.g. temporary
files) because they have a dependence with the driver itself.

This change basically makes all the libraries in the compiler static
ones and then they are gathered into a single libmercurium shared
library which is the one the phases should only need to link again. As
mentioned above this is currently not possible because a phase may end
using features of the driver (so these should be extracted first). That
said due to the way shared libraries work in Linux this is not a
problem at the moment.

A few phases in analysis relied on global variables, unfortunately
their construction is not possible before main, so functions that
construct the wanted values are used instead. This is certainly less
efficient but avoids the pitfall of building global data too early
in the pipeline.